requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
DataRecorder==3.6.2
datasets==3.6.0
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
DownloadKit==2.0.7
DrissionPage==4.1.0.17
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.3
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
kafka==1.3.5
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
ollama==0.4.5
openai==1.59.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
psutil==6.1.1
pyarrow==20.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rouge_score==0.1.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.27.1
sentence-transformers==4.1.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
sympy==1.13.1
threadpoolctl==3.6.0
tldextract==5.1.3
tokenizers==0.21.1
torch==2.5.1
tqdm==4.67.1
transformers==4.49.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
venn==0.1.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
